Added registrar landing page.
Adds a proper landing page to the registrar role with buttons to navigate to reports and the map. Adds a database connection check as well as a placeholder for pending reports which needs to be changed when the needed functionality is added. Made authorization changes in dataform so that registrars can only view the map but cant add markers. Made it so that all users that are in the registrars role are redirected to the correct landing page on login as well as when pressing the home button in the navbar and the Digital Obstacle Report header.